donut_invoice_model/Lec_4/requirements.txt
# OCR和文档处理
paddlepaddle>=2.4.0
paddleocr>=2.7.0
pdf2image>=1.16.0
Pillow>=9.0.0
opencv-python>=4.5.0
pytesseract>=0.3.10

# 机器学习和深度学习
torch>=1.9.0
torchvision>=0.10.0
transformers>=4.21.0
datasets>=2.0.0
Pillow>=8.3.0
numpy>=1.21.0
scipy>=1.7.0
scikit-learn>=1.0.0
tqdm>=4.62.0
accelerate>=0.12.0
sentencepiece>=0.1.96
protobuf>=3.19.0
requests>=2.26.0

# 数据处理
pandas>=1.3.0

# 其他工具
tqdm>=4.62.0
requests>=2.25.0